starting to extract a LicenseDownloader
