+++ src/KeyHub.Client/KeyHub.Client.csproj
@@ -42,6 +42,7 @@
   </ItemGroup>
   <ItemGroup>
     <Compile Include="LicenseVerifier.cs" />
+    <Compile Include="LicenseDownloader.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
   </ItemGroup>
   <ItemGroup>
