+++ src/KeyHub.Integration.Tests/issues.txt
@@ -1,10 +1,5 @@
 ﻿
-
-
-Nathanael will be out Wednesday (out of office, but maybe on email at end of day), available Thursday afternoon.
-If get stuck, move onto the licensing code
-  (like https://github.com/imazen/keyhub/issues/36)
-  
+KeyHub.Client shouldn't have a dependency on ImageResizer  
 
 
 How to repro 50 character email problem?
